move encoders to mol-io
