+++ src/mol-io/writer/mol2.ts
@@ -5,7 +5,7 @@
  */
 
 import { Mol2Encoder } from './mol2/encoder';
-import { Encoder } from '../../../mol-io/writer/cif/encoder';
+import { Encoder } from './cif/encoder';
 
 export namespace Mol2Writer {
     export interface EncoderParams {
